receipt_ocr/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
easyocr==1.7.1
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
gunicorn==21.2.0
torch==2.1.0
torchvision==0.16.0